+++ src/Symfony/Component/DependencyInjection/Tests/Fixtures/config/defaults.php
@@ -13,9 +13,9 @@ return function (ContainerConfigurator $c) {
         ->autoconfigure()
         ->autowire()
         ->tag('t', ['a' => 'b'])
-        ->bind(Foo::class, ref('bar'))
+        ->bind(Foo::class, service('bar'))
         ->public();
 
-    $s->set(Foo::class)->args([ref('bar')])->public();
+    $s->set(Foo::class)->args([service('bar')])->public();
     $s->set('bar', Foo::class)->call('setFoo')->autoconfigure(false);
 };